Cleaning, preparing for playlist implementation
Cleaning up some comments, adding some new functions and classes to prepare for the new way playlists will be handled. Essentially a function call to generate a playlist request will be required now. This is to allow the user to see the items in the playlist, and that playlist request will be provided to the playlist downloader which will then perform the rest of the actions